SBERDOMA-795 Client hooks separated
Now using separate react hook for creating Importer (common for all usages)
and hook for generating Importer props
