+++ apps/condo/domains/property/hooks/useImporterFunctions.tsx
@@ -1,20 +1,44 @@
-import { useCallback, useRef, useState } from 'react'
 import { useOrganization } from '@core/next/organization'
 import { useApolloClient } from '@core/next/apollo'
 import { useAddressApi } from '../../common/components/AddressApi'
 import get from 'lodash/get'
 import { Property } from '../utils/clientSchema'
 import { searchProperty } from '@condo/domains/ticket/utils/clientSchema/search'
-import { TableRow, ProcessedRow, Importer, ColumnInfo } from '@condo/domains/common/utils/Importer'
+import { TableRow, ProcessedRow, ColumnInfo } from '@condo/domains/common/utils/Importer'
 import { MapEdit, MapTypesList } from '../components/panels/Builder/MapConstructor'
 
-export const useImporter = (onFinish, onError) => {
-    const [progress, setProgress] = useState(0)
-    const [error, setError] = useState(null)
-    const [isImported, setIsImported] = useState(false)
+const createPropertyUnitsMap = (units, sections, floors) => {
+    const unitsOnFloor = Math.floor(units / (floors * sections))
+    if (!unitsOnFloor) {
+        return
+    }
+
+    const propertyUnitsMap = {
+        dv: 1,
+        type: MapTypesList.Building,
+        sections: [],
+    }
+
+    const mapEditor = new MapEdit(propertyUnitsMap)
+
+    for (let currentSection = 0; currentSection < sections; currentSection++) {
+        const name = `№${currentSection + 1}`
+        mapEditor.addSection({
+            name,
+            unitsOnFloor,
+            minFloor: 1,
+            maxFloor: floors,
+        })
+    }
+
+    // TODO(Dimitree): thing about rest flats
+    return mapEditor.getMap()
+}
+
+// eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
+export const useImporterFunctions = () => {
     const userOrganization = useOrganization()
     const client = useApolloClient()
-    const importer = useRef(null)
     const { addressApi } = useAddressApi()
 
     const userOrganizationId = get(userOrganization, ['organization', 'id'])
@@ -55,36 +79,6 @@ export const useImporter = (onFinish, onError) => {
             })
     }
 
-    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-    // @ts-ignore
-    const createPropertyUnitsMap = (units, sections, floors) => {
-        const unitsOnFloor = Math.floor(units / (floors * sections))
-        if (!unitsOnFloor) {
-            return
-        }
-
-        const propertyUnitsMap = {
-            dv: 1,
-            type: MapTypesList.Building,
-            sections: [],
-        }
-
-        const mapEditor = new MapEdit(propertyUnitsMap)
-
-        for (let currentSection = 0; currentSection < sections; currentSection++) {
-            const name = `№${currentSection + 1}`
-            mapEditor.addSection({
-                name,
-                unitsOnFloor,
-                minFloor: 1,
-                maxFloor: floors,
-            })
-        }
-
-        // TODO(Dimitree): thing about rest flats
-        return mapEditor.getMap()
-    }
-
     const propertyCreator = (row: ProcessedRow | null) => {
         if (!row) return Promise.resolve()
         const [, units, sections, floors] = row.row
@@ -102,36 +96,5 @@ export const useImporter = (onFinish, onError) => {
         })
     }
 
-    const importData = useCallback((data) => {
-        importer.current = null
-        // reset hook state
-        setIsImported(false)
-        setError(null)
-        setProgress(0)
-
-        // @ts-ignore
-        importer.current = new Importer(columns, propertyNormalizer, propertyValidator, propertyCreator)
-        importer.current.onProgressUpdate(setProgress)
-        importer.current.onError((e) => {
-            importer.current = null
-            setError(e)
-
-            onError()
-        })
-        importer.current.onFinish(() => {
-            importer.current = null
-            setIsImported(true)
-
-            onFinish()
-        })
-        importer.current.import(data)
-    }, [])
-
-    const breakImport = () => {
-        if (importer) {
-            importer.current.break()
-        }
-    }
-
-    return [importData, progress, error, isImported, breakImport]
+    return [columns, propertyNormalizer, propertyValidator, propertyCreator]
 }
